Add new Monte Carlo approach
The problem of finding a valid solution for a given crossword puzzle is
NP complete, so using a random educated guess is better than trying to
find an optimal solution that works every time.

This new approach uses the following method:

1. Find all possible matchign candidates for all word slots
2. Find the word slot with least matching candidates
3. Select a candidate randomly and check if the board goes to an illegal
   state after selection
3a. If state is legal, continue loop from 2. until all words are filled
3b. If state is illegal, reset everything and start again from 1.

This could be further improved with backtracking, but it adds a ton of
complexity too.

The code is written in such way that it avoids copying and it should be
as performant as possible. It contains multiple caching optimizations to
improve main loop speed by precalculating lookup tables etc.

Since we only need a single answer for a problem, this single-threaded
program can be run in multiple instances to parallelize it to use all
cores, see `yarn start:parallel` script.

Further optimizations:

* Backtracking
* Rewrite in Rust